pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tso-robotics-sockets"
version = "0.1.0"
description = "Lightweight ZMQ socket communication for robotics inference servers"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Lorenzo Mazza", email = "lorenzo.mazza@nct-dresden.de"},
    {name = "Ariel Rodriguez", email = "ariel.rodriguezjimenez@nct-dresden.de"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Networking",
]
dependencies = [
    "pyzmq",
    "numpy",
    "opencv-python-headless",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
]

[project.urls]
Repository = "https://github.com/Lorenzo-Mazza/tso_robotics_sockets"

[tool.hatch.build.targets.wheel]
packages = ["src/tso_robotics_sockets"]

[tool.pytest.ini_options]
markers = [
    "unit: Fast unit tests with no external dependencies",
    "integration: Socket tests requiring ZMQ transport",
]